requirements.txt
pyfunctional==1.2.0
pyinflect==0.5.0